+++ src/quad3d_editor_update_lighting.cpp
@@ -47,37 +47,37 @@ void Quad3dEditor::_updateLightingMenu()
 
 		if((_gui->getOverlay()->getValueFormId() == "colorR") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent());
+			const auto content = static_cast<float>(Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent()));
 
-			_fe3d->quad3d_setColor(_currentQuadId, (static_cast<float>(content) / 255.0f));
+			_fe3d->quad3d_setColor(_currentQuadId, (content / 255.0f));
 
 			_gui->getOverlay()->openValueForm("colorG", "G", (color.g * 255.0f), fvec2(0.0f, 0.1f), 5, false, true, false);
 		}
 		if((_gui->getOverlay()->getValueFormId() == "colorG") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent());
+			const auto content = static_cast<float>(Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent()));
 
-			_fe3d->quad3d_setColor(_currentQuadId, (static_cast<float>(content) / 255.0f));
+			_fe3d->quad3d_setColor(_currentQuadId, (content / 255.0f));
 
 			_gui->getOverlay()->openValueForm("colorB", "B", (color.b * 255.0f), fvec2(0.25f, 0.1f), 5, false, true, false);
 		}
 		if((_gui->getOverlay()->getValueFormId() == "colorB") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent());
+			const auto content = static_cast<float>(Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent()));
 
-			_fe3d->quad3d_setColor(_currentQuadId, (static_cast<float>(content) / 255.0f));
+			_fe3d->quad3d_setColor(_currentQuadId, (content / 255.0f));
 		}
 		if((_gui->getOverlay()->getValueFormId() == "lightness") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent());
+			const auto content = static_cast<float>(Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent()));
 
-			_fe3d->quad3d_setLightness(_currentQuadId, (static_cast<float>(content) / 100.0f));
+			_fe3d->quad3d_setLightness(_currentQuadId, (content / 100.0f));
 		}
 		if((_gui->getOverlay()->getValueFormId() == "emissionIntensity") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent());
+			const auto content = static_cast<float>(Tools::parseSignedInteger(_gui->getOverlay()->getValueFormContent()));
 
-			_fe3d->quad3d_setEmissionIntensity(_currentQuadId, (static_cast<float>(content) / 100.0f));
+			_fe3d->quad3d_setEmissionIntensity(_currentQuadId, (content / 100.0f));
 		}
 
 		screen->getButton("isBright")->setTextContent(isBright ? "Bright: ON" : "Bright: OFF");
